Move custom calendar widget javascript implementation out of the core.js and into a separate file
In core/maintained packages, this widget is used only by template::widget::textdate (which in turn is used nowhere). For this widget, we can include the relevant code on-demand.
Direct usages found in unsupported packages have been adapted by adding the extra requirement on the fly.

It is unclear if we should keep maintaining a calendar widget in our codebase, however, as the current one still works (kudos to the original author), we do not deprecate it for the time being.
It is also true that a custom widget provides a way to customize the look and feel of date fields, which is otherwise completely delegated to the browser.
Various packages, including the core, provide custom css styling for the current widget.